go-fips-1.22: drop the md5 test
I expected that an md5 test would fail in a fips enviroment
failure due to the algoritm being restricted.

Currently the 'md5' executable does not have a C dependency (ldd shows
'not a static executable') and no attempts are made to open libcrypto.
This happens independent of a fips provider.

So for now, the test is only verifying what I understand to be
broken behavior.  I'd rather remove that test than confuse someone
later on if it starts to fail.