go-fips-1.22: declare boringcrypto GOEXPERIMENT
Our extra patch makes boringcrypto codepaths unconditional. However,
many external projects are sensitive to this experiment and build
tag. Build our fips toolchain with boringcrypto GOEXPERIMENT
explicitely enabled. This results in boringcrypto being listed in the
output of `go version -m path/to/binary` which are built with this
toolchain. This aids with positive indicator to identify if a given go
binary was compiled in FIPS mode or not.

Signed-off-by: Dimitri John Ledkov <dimitri.ledkov@chainguard.dev>