py3-hashin - install in a virtual env.
Install py3-hashin in its own virtual environment.
Link it into /usr/bin so it can be used without knowledge.

This will allow the py3-hashin package to be installed without
conflicting with other packages.

This generally *breaks* the use of it as a library.
ie 'import hashin'

For the moment that is OK as py3-hashin is only used as a requirement
for renovate, and renovate invokes 'hashin' executable.

Signed-off-by: Scott Moser <scott.moser@chainguard.dev>