Drop py3-azure-*, py3-adal, py3-ms* packages
These packages were added in an effort to provide 'az' (azure cli
client).  That package was ultimately added in wolfi as
az.yaml using 'pip install' via under
https://github.com/wolfi-dev/os/pull/10514

End result: These are not used anywhere and not needed.
If they are needed in the future they can be resurrected from
git history.

Signed-off-by: Scott Moser <smoser@chainguard.dev>