go-msft: Enable coverage under fips mode; whilst blocking md5 usage under FIPS (#8171)
Patch exception reason:
- Improves FIPS support, by blocking unapproved MD5 (RC4 DES 3DES) algorithm
- Does not regress upstream standard features support (coverage)

Not sure what is better - patching in a copy of src/crypto/md5, or do `cp -r src/crypto/md5 src/internal/coverage/md5` and then only have the patch that changes imports.

The customer issue is that "but it is being blocked by go-msft image provided by MS on Mariner / Azure Linux container images" and they observed that it wasn't on Chainguard.

This is due to the checks in this go-msft toolchain in combination with openssl fips 1.1.1 shipped in Mariner image. Our openssl FIPS doesn't have these obsolete algorithms and reports them as not-available. openssl fips 1.1.1 reports them as available, but then causes them to fail at runtime. Hence this patch to remove this discrepancy.

Refactoring this and submitting it upstream with extra support for all other backends too, which we do not use in wofli:
- https://github.com/microsoft/go/pull/1346
- https://github.com/microsoft/go/pull/1348

Thus once those make it into microsoft/go release, we will be able to drop these extra patches.