bcfips-policy: make maven just work (#8281)
Very common complaint against jdk-fips images is that "maven does not work".
Well make maven just work.

This is very similar to keytool, one has to set truststore type &
module-path to be able to use FIPS truststore type, and find
module-path to bouncycastle fips modules.

Add test-case that builds maven when JDK is in FIPS approved only
mode, with full test suite. It shows that maven is using FIPS-approved
cryptography only.

Also demonstrate that it is possible to build zookeeper when JDK is in
FIPS-approved only mode. This opens up a door to run upstream test
suites in FIPS mode.