datawire-envoy-fips: Cherrypick googleurl fix (#8547)
There was an upstream breakage of envoy from a dependency accidentally
getting deleted from a storage bucket. The dependency was placed back,
but has a new SHA due to an epoch change.

This cherrypicks the fix from main in order to fix the sha.
We're doing this here to get CI to turn green for this package since it
is causing alerts.