openssl-config-fipshardened: add openssl-fips-test by default (#8606)
openssl-fips-test is a 20kb binary that is extremely useful for FIPS
compliance and auditing checks as it enables executing it as an
entrypoint, or on-demand in a running container to verify that openssl
is running in fips mode. And it is able to do so without shell or
/usr/bin/openssl binaries present.

This is a soft-launch of this feature, with academy documentation to
follow.

Note this has a side-effect, if a container didn't previously have
libcrypto.so/libssl.so they will now be pulled in. Previously
openssl-config-fipshardened only pulled in fips.so, and expected for
libcrypto.so/libssl.so to already be in the dependencies or be pulled
in at a later time somehow.