gotenberg - add a wrapper to set environment, improve test
Gotenberg needs a number of environment variables set to run.
Its silly to have 2 copies of those environment variables,
one in the test and one in the image config.

With this change we can actually drop the list in the image config.

Its better to have one copy and have it in the package, so that
changes to the package (like the recent update that needed)
PDFCPU_BIN_PATH does not need to be copied to the image.

Signed-off-by: Scott Moser <scott.moser@chainguard.dev>