feat: Add Open Liberty with SCC (#9978)
Migrates Open Liberty to enterprise as we are now generating a class
cache to be used by OpenJDK with OpenJ9. Also creates a defaultServer
that is useable from the get go inline with upstream

The packaging for this is an absolute... dumpster fire isn't even good
enough. This is the packaging equivalent of a burrito from Cafe Rio

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>