cilium-fips-1.16: make JSON version available (#10311)
The upstream code in cilium/cilium assumes that `go version` has no
space. However, this is untrue when GO_EXPERIMENT is set.

Our go-fips has `GO_EXPERIMENT=X:systemcrypto`, and does not conform to
this assumption. We should send a patch upstream, but in the meantime,
let's fix it in enterprise-packages in order to unpin the cilium
connectivity test (newer tests need Cilium structured versions).

Signed-off-by: Nghia Tran <tcnghia@gmail.com>