Package guest agents for GCE
The packaging here is adapted from the packaging subdirectory in the
upstream repos.  The runtime dependencies (other than dracut) are
included.

selinux contents are included because of the policycoreutils requirement
of google-guest-oslogin.   All of those packages are from the
same repo and those tags all point to the same commit. Ideally we'd just
use one file and they'd all at the same time. Unfortunately because of
dependencies between packages it's annoying to build them as a
subpackages of a single package. Other distros build them from the
separate release tarballs.