linux - add an /etc/fstab for rootfs as read-write (#12875)
This is probably not the final resting place for this,
but for now we need to tell systemd that it should mount / as rw
or it doesn't get mounted rw.

This may end up in another package or something at some point.
Other options for telling systemd to mount rw:
a. boot with kernel command line 'rw'
b. add a systemd mount unit