Support reading keys from DMI oem strings and kernel cmdline.
Example of how to boot qemu such that it would include

   $ pubkey="ssh-ed25519 AAAAC3Nza.....0nUYz smoser@smdev"
   $ qemu-system-x86_64 ....
      -smbios type=1,serial=chainguard-fa310a0da8e844d7b9f55bc10741f5dc
      -smbios "type=11,value=chainguard.dev/v1/ssh-pubkey=$pubkey"

If you are able to modify the kernel command line, then you can
also do it on the cmdline like this:
also pass the key in that way with:

   $ pubkey="ssh-ed25519 AAAAC3Nza.....0nUYz smoser@smdev"
   $ b64pub=$(echo "$pubkey" | base64)
   $ qemu-system-x86_64 ....
      -smbios type=1,serial=chainguard-fa310a0da8e844d7b9f55bc10741f5dc
      -append "... sshkey=$b64pub"

The string 'fa310a0da8e844d7b9f55bc10741f5dc' is just something
I made to be unique id for a given "version" of the "platform".
This way we can detect that in the guest and make decisions
based on it.