Add chainguard-baselayout as kernel runtime dep. (#13162)
Dustin noticed that our generic.json prints Wolfi at boot, and this was because we weren't including `chainguard-baselayout`.

Including `chainguard-baselayout` explicitly makes it print Chainguard instead of Wolfi, but we want to make sure that our VMs always carry the Chainguard branding, so I am adding this runtime dependency so we can't forget.

Signed-off-by: Matt Moore <mattmoor@chainguard.dev>