cloud-ssh-keys-fetcher - workaround 2 issues
It seems HOME isn't set in the systemd environment on our systems
right now, so hackily figure out HOME. This is only really needed
for root user as that is where keys are imported with the unit.

Other issue here is to update PATH to have /bin:/sbin as we need
it for some programs there.