google-guest-agent-manager: New package for gce agents
This package is a single binary from a different version of the
guest-agent.  You can find more detail at the following links:

We package guest-agent as google-guest-agent[1]
as google-guest agent. In the specfile included[2] in geust-agent it
mentions building with extras. according to the overview[3] that extra
piece is coming from the google-guest-agent package which is what we
have here.

1. https://github.com/GoogleCloudPlatform/guest-agent/tree/main
2. https://github.com/GoogleCloudPlatform/guest-agent/blob/main/packaging/google-guest-agent.spec
3. https://github.com/GoogleCloudPlatform/guest-agent/tree/main?tab=readme-ov-file#overview