fix build-images.yaml (#40)
https://github.com/chainguard-dev/wolfi-vm/actions/runs/13000836492

```
The workflow is not valid. .github/workflows/build-images.yaml (Line: 65, Col: 13): The identifier 'build images' is invalid. IDs may only contain alphanumeric characters, '_', and '-'. IDs must start with a letter or '_' and and must be less than 100 characters.
```

But really all of these `id`s were unnecessary anyway, and could just
happen in the same step.

Signed-off-by: Jason Hall <jason@chainguard.dev>