Make debug disk have 'rw' if the non-debug version has 'rw'. (#43)
linux kernel root= flip-flopped on presense of 'rw'. the debug version
didn't get updated. That meant that 'make debug-disk-generic' would
_not_ have 'rw' on it while the regular version would.

This is hacky, but we search for 'rw' from the main and then adjust the
debug entry to have it.

Long term solution is to have a debug package (or metadata package) that
does this and so we don't fish around and hackily build content during a
debug build process but rather just include debug packages.