write configs as YAML instead of JSON (#47)
The decision to write configs as JSON seems to stem from the history
that they were extracted from OCI image attestations, which are in JSON.

New hand-authored configs are in JSON for no reason (AFAIK), and we
don't seem to be doing much with attestation-generated configs right
now.

In any case, YAMLifying JSON is easy with `yq`, and the result is
something that (IMO) is easier to read, easier to write, and easier to
lint. So let's use YAML.

Feel free to reject this PR if any of the previous sentences are
incorrect or controversial.

---------

Signed-off-by: Jason Hall <jason@chainguard.dev>