ec2-instance-connect: fix it to use cert bundle
ec2-instance-connect does a lot of certificate parsing, by default it
expects a cert directory with individual .pem files, but we only ship
a single certificate bundle by default. Switch the scripts to use the
certificate bundle.

Also capture a realistic example, and execute it at build time, to
help verify that ec2-instance-connect works, as it is very difficult
to debug it life as the keys are short lived and are revoked in
seconds.

Add required runtime dependencies, requires usr-merge & gnu tooling.

Requires faketime:
- https://github.com/wolfi-dev/os/pull/40898