feat(request-3790): move jupyter-base-notebook to custom pkg (#13771)
This change comes from the fact that the jupyter-base-notebook
does not align to upstream base-notebook, missing the jupyterhub-singluser
conda package, which maps to the juypyterhub pkg on pypi, but
the image agreed with the customer is still the one that includes
this current package.
The result is a custom package with a custom image, while the
future possible jupyter-base-notebook should align to the same
bill of material the upstream has in quay.io/jupyter/base-notebook.

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>