Improve dependencies listed in Makefile (#83)
There are several changes here to get the build arch (qemu-system-XXXX)
to be able to be specified. That allows me to debug a mac system
(mostly). Some of the changes getting qemu args are needed anyway in
order to be able to _run_ a different arch than your cpu. Ie, we want to
be able to qemu-system-aarch64 to inspect a aarch64 disk. As it was
previously, qemuArgs was only available for the cpu that you were
running on.

Other things to get there:
1. dependencies in the make file are now specified for everything that
we build.
2. add 'disk-<name>' and 'disk-debug-<name>' targets for all config
files. (this tab completes too)
3. add 'fetch' subcommand for getting artifacts much like the
grab-pkg-artifact used to do.