fix(karpenter-fips): build with go-fips-md5 toolchain (#13848)
Karpenter autoscaler leverages AWS SQS [1] to manage autoscaling
related messages in a event-driven fashion. It uses the AWS SDK
to communicate with the service. The middleware uses crypto/md5.Sum()
to validate [2] the message returned by the SQS API by calculating
the md5 checksum [3] on the input message body to the ReceiveMessage
operation [4].

1. https://github.com/aws/karpenter-provider-aws/blob/cf530be3573ae9865183ae204cf33bfa3d0fdcae/pkg/providers/sqs/sqs.go#L70
2. https://github.com/aws/aws-sdk-go-v2/blob/68c46490c32137ae1449ff3a33e88337b814903d/service/sqs/cust_checksum_validation.go#L109
3. https://github.com/aws/aws-sdk-go-v2/blob/68c46490c32137ae1449ff3a33e88337b814903d/service/sqs/cust_checksum_validation.go#L196
4. https://github.com/aws/aws-sdk-go-v2/blob/68c46490c32137ae1449ff3a33e88337b814903d/service/sqs/api_op_ReceiveMessage.go#L366

Signed-off-by: Massimiliano Giovagnoli <massimiliano.giovagnoli@chainguard.dev>