build ingress-controller-1.11 again from otel-contrib (#14528)
* build ingress-controller-1.11 again from otel-contrib

if we always use the latest commit to build then it fails and on every
new commit we get unexpected commit error so setting depth to -1 to do a
hard checkout.

there was a regression introduced upstream where they remove
opentelemetry_config so using this commit should fix it.

upstream is using the same commit (ref:
https://github.com/kubernetes/ingress-nginx/blob/ad75df8a5193f0d8c1a317ef6b526f4039c5b157/images/nginx/rootfs/build.sh#L538)

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>

* remove unreconized entries from nginx.conf

these were introduced in main and still not available on the commit
we're checking so removing this.

not adding opentelemetry_config in here given that's specific to the
oldbranch and will on update of the package.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>

* add opentelemetry specific keys to nginx.conf

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>

---------

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>