tritonserver-backend-vllm-24.04: Bump vllm to latest version (0.7.2)
I had bumped to 0.7.0, not to the latest (0.7.2) because 0.7.0 addresses
a CVE, and I figured it would be good to test this bump before going
all the way to the latest. Well, bumping to 0.7.0 builds OK, but fails
to test because the video modality tries to load a python module we don't
include:

2025/02/17 17:06:32 INFO running step "Verify vllm installation"
[...]
2025/02/17 17:06:34 WARN   File "/opt/tritonserver/venv/lib/python3.10/site-packages/vllm/multimodal/utils.py", line 20, in <module>
2025/02/17 17:06:34 WARN     from .video import VideoMediaIO
2025/02/17 17:06:34 WARN   File "/opt/tritonserver/venv/lib/python3.10/site-packages/vllm/multimodal/video.py", line 7, in <module>
2025/02/17 17:06:34 WARN     import cv2
2025/02/17 17:06:34 WARN ModuleNotFoundError: No module named 'cv2'

0.7.2 made this import optional:
  https://github.com/vllm-project/vllm/commit/20579c0fae1757c9da9fc35a69960563186f3036

Let's bump to 0.7.2 and avoid resolving this import because our goal here
is to patch a CVE, not to add new functionality to this version stream.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>