tritonserver-backend-vllm-24.04: Drop usage of /usr/bin/python3
This package does not require /usr/bin/python3 -> python3.10.
The image may - and currently does. python-3.10-base will
automatically be added as a dep by melange.

We should also ensure we are building with /usr/bin/python3.10,
and not relying on a python3 there either. Let's replace the
build-dep on python-3.10-dev with py3.10-build-base-dev and
remove the dependencies this package already brings in.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>