github: do not fail grype scan (#157)
In packages & images whilst grype scan is perfromed, it is never
blocking. Also currently grype fails to parse syft.sbom.json as it
appears that the generated syft.sbom.json is invalid and/or
incomplete.

Ensure that daily image building continious until grype scanning is
fixed:

> [0016] ERROR failed to catalog: unable to decode sbom: unable to
decode syft-json document: unable to find source metadata type=""

Note there is likely incompatibility with new grype, as there was a new
release of it last week:
- https://github.com/anchore/grype/releases/tag/v0.89.0