cloud-ssh-keys-fetcher: rewrite as AuthorizedKeysCommand
The benefits of AuthorizedKeysCommand is that it can lookup keys on
demand, when a connection is being established. It can do so for any
account (for better or worse). And it can echo them to stdout, without
need to write them out anywhere.

Plus it can integrate with ec2-instance-connect to allow connecting
either via metadata key or the user key.

Note that systemd supports native secrets settings via the smbios &
kernel cmdline for root password & root ssh key. Thus it is redundant
to implement that. However I have not tested which components of
systemd are required for that to run.

As we are about to launch the product, it is best to focus on just
in-public cloud usecases only.