py3.10-torch-2.6-cuda-12.4: Add sm_75 to TORCH_CUDA_ARCH_LIST
This is the architecture of the Tesla T4, which is commonly
available in public clouds.

It is also supported by both of our upstream reference images
(as described in the comment block).

Without this, our tritonserver-backend-vllm image spews a ton
of these errors on startup:

```
FATAL: kernel `fmha_cutlassF_f16_aligned_64x64_rf_sm75` is for sm75-sm80, but was built for sm70
FATAL: kernel `fmha_cutlassF_f16_aligned_64x64_rf_sm75` is for sm75-sm80, but was built for sm70
```

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>