cloud-ssh-keys-fetcher: ignore eic binary failure
If there are no keys pushed to eic within the last 60s executing the
binary will error out. This is undesired behaviour, as that prevents
trusting already ec2dump fetched keys.

A performance optimisation and improvement would be to check the
passed in fingerprint, and exit as soon as it is found in the ec2dump
keys or the eic.