chore: Withdraw Hadoop FIPS and gRPC Java FIPS (#18645)
We currently create a local maven repository that consists of all
dependencies for Hadoop and gRPC used by Spark at buildtime. This also
means we carry all of the CVEs associated with all the JARs we vendor.
This is not practical to maintain for sustaining. Let's build everything
inline for now instead

Signed-off-by: RJ Sampson <rj.sampson@chainguard.dev>