Remove unused configs from the production configs directory (#191)
Remove currently unused image configs, that were old chainit
prototypes, not currently in active use; not build daily; not tested;
not published; do not comply with expected yam formatting.

The functionality similar to curl/crane is still excercised daily as
the "builder.yaml" for image building, so removing these configs does
not remove/regress support for such single-purpose appliances, as the
builder remains an example of that.