Expand wolfi-base metapackage into dependencies (#192)
wolfi-base is a metapackage that pulls in apk-tools, busybox, and
wolfi-keys.

For many of our images we may want apk-tools & wolfi-keys, but not
busybox.

Expanding wolfi-base into its dependencies, shows a clearer picture of
the package configuration in our images. This also reduces manifest of
installed packages by one.

Also run yam on all the configs.