feat(tempo-2.3.yaml): Deprecate tempo 2.3 in favour of latest `tempo` package
While investigating a couple of tempo-2.3 CVEs https://github.com/chainguard-dev/enterprise-packages/pull/16730 I noted that the latest tempo
image is using this once off tempo-2.3 package which does not appear to be supported upstream [1] and last received an update in November 2023 and is now vulnerable to CVEs.

There is a tempo package @ version 2.7 which is supported upstream.

[1] https://github.com/grafana/tempo/releases

Signed-off-by: philroche <phil.roche@chainguard.dev>