Fixup some naming and architecture issues with azure uploading (#215)
.github/workflows: Prevent double hyphen in Azure


azure-image-upload: Resepect the arch setting
    
The script seems to be ignoring the arch flag. It seems like archin is
always
set to empty then checked if it's empty then arch is set to uname. I
think this
explains why some of my images have been misuploaded.  I'd rather this
require the arch. There's no real connection between the host you're
running on and the image it should upload. If the arch is going to keep
being inferred then it should we should integrate it into the path for
uploads and take in just the name and assemble the path like:
`./output/$arch/$name.raw`

---------

Co-authored-by: Justin Vreeland <justin.vreeland@chainguard.com>