repmgr-*: Update bitnami compat commit (#20162)
This commit updates the pinned commit fetched for the bitnami compat
subpackages, strictly speaking only repmgr-14 and repmgr-15 are affected
but for consistency we'll update all of them.

The prior pinned commit pinned to early and missed out on the next 2
commits which specifically target the repmgr-14 and repmgr-15 versions.