Stop updating bc-fips 1.x packages
bc-fips 1.x supported ended in July 2024, and grace period finished in
January 2025.

We are about to start upgrade of bc-fips 2.0 to 2.1.

All of these are no longer in use (images for the -140-2 no longer
built), or they have converted to using the policy files.

Remove subpackages, or full yaml files which are no longer in use - or
more specifically no longer need to be updated.

The apk remain published.

This is cleanup to finish transition of 1.x.

Also see:
- https://github.com/chainguard-images/images-private/pull/8554
- https://github.com/chainguard-images/images-private/pull/8562