bouncycastle-fips: bump to 2.1.0 release
Bump Bouncy Castle FIPS packages to the 2.1.0 release as this
has now been certified.

This is a minor release upgrade with no breaking changes.

Drop jmail, mail and pg packages as runtime dependencies for
the core provider; they are not mandatory and having both jmail
and mail installed at the same time causes issues with module
loading in Java.

Refresh bcfip-policy-140-3 to require ~2.1 release of
Bouncy Castle FIPS (and refactor package a bit to make that
easier in the future).

Signed-off-by: James Page <james.page@chainguard.dev>