nvidia-cuda-cudart-*: Add versioned symlink to libcuda.so
libcuda.so.1 is tightly coupled to the GPU driver running on the
host, so it needs to be passed in to the containers that use it.
Since this library isn't present during package testing, our
`ldd-check` test fails for binaries that link against it.

We can set LD_LIBRARY_PATH to the stub library path instead, since
the stub will satisfy ldd's check. But for that to work, we need
to have a versioned library, not just the .so.

Use ldconfig to create a versioned symlink, currently:
  libcuda.so.1 -> libcuda.so

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>