Add helpers/install-deps script for installing deps.
This may well be over-engineered at the moment.
The immediate need is for wolfi-vm to use this repo during
its c-i and not duplicate knowledge of what dependencies
are needed.

So the plan here is to have wolfi-vm call helpers/install-deps

Later this same 'install-deps' can be OS aware and install
packages necessary if it is running inside wolfi or another os.