rancher-webhook-0.4: remove bump to k8s deps (#22334)
This part of the code never worked, and we were always using the k8s
version 1.28.6 from upstream.
When we attempted to fix the code, any bump in version will always lead
to build failures due to API breakages between versions.
We will need to rely on upstream to bump the dependencies and fix the
code when doing so.