ct-manylinux-2.28: the cross native toolchain for Chainguard Libraries
This is a partial native toolchain, built with lower:
- march
- mtune
- glibc

Such that sysroot produced libraries are bootstrapped for lower glibc
ABI, with support for C, C++ and Fortran.

Some aspects of the toolchain are still very up to date, for example
using latest linux kernel headers to ensure access to latest kernel
features on opt-in feature guarded basis.

It is integrated automatically via openssf-compiler-options, such that
mere addition of this package to the build environment should lower
the ABI of the package being built.