Build py3.*-torch-2.6-cuda-12.4 from single YAML file and refactor
It turns out that the libtorch C++ library is useful without any of
the python bits. So here we build that separately, in its own apk,
and then build the python-specific bits for each python version against
it.

Since we do this by building whls for each stage, that means we are
building 5 whls from this source. 1 for libtorch, and 1 for each of
the 4 supported python versions. The `python -m build` process we were
using would build an sdist each time, which takes a lot of time for
a package this size. So here we build 1 sdist to start, and then
use `python3 -m pip install` to build binaries from it at each stage.

We supply a py3-supported- package as a courtesy so that dependent
libraries can pull in the libs for all python supported versions at once.
This means they need to be co-installable. To make that possible, we
need to split the /usr/bin/ into separate packages. This means that
the new py3.x-torch-2.6-cuda-12.4 is *not* a 1:1 replacement for the
pre-merged py3.x-torch-2.6-cuda-12.4. You want py3.x-torch-2.6-cuda-12.4-bin
instead.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>