py3.*-torch-2.6-cuda-12.4: provide ${{package.name}}-bin
We're working on merging pytorch packaging so that all python
versions are produced from the same YAML, sharing a single
C++ libtorch build. We want to be able to co-install all of
these py3.x-* builds of torch so that we can use them as build
dependencies for torchvision packages that will also build
all python versions from a single YAML. Doing so requires splitting
the pytorch commands into separate -bin packages, which images will
need to pull in to avoid losing the commands.

If an image were built between the time `-bin` packages are split out
but before the image configs are updated to pull them in, those
images would be published without commands.

To avoid this race, let's start by adding `-bin` as a virtual package
in all of the non-merged packages. Then we can switch images over to using
-bin as a no-op change. After that, introducing the split-out `-bin`
packages should seemlessly complete the transition.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>