grafana-alloy-fips: build with latest go-fips-md5
Upstream have moved to 1.24, so follow that move.

This also adds a build dependency on nofips-node, because go-fips-1.24
configures the system to completely reject MD5, which causes problems
for the many NPM packages which still use it. `nofips-node` is a wrapper
which overrides the system-wide configuration to allow the NPM parts of
the build to continue functioning.