adding nofips-node into build environment
The go-fips-1.24 package differs from previous go-fips packages in that it configures the system's OpenSSL to reject non-FIPS compliant usage. We've seen this cause problems for packages which also have a Node component, because of NPM packages still use MD5. To allow us to continue configuring the system correctly while allowing these packages to build.


Signed-off-by: Debasish Biswas <debasishbsws.dev@gmail.com>