Add aws test helper.
The test input takes 'awspub' output as its input format.
See the example in Usage for what those look like.
The arch is sniffed from the image name.

Other things here is a mod to the primary helper script to
shift off cloud and resultsd, and then the modification to
the azure and gcp helpers to address that (I do not think
I actually had to do that, as they are currently just
getting additional arguments that they're not using).

One thing to note is that we're currently not waiting around
for console logs to show up.  So I think that means we're
probably getting empty output most of the time.

Other thing to note is that instance type is ignored/"fixed"
for aarch64.  if t3.medium (the default) then we just change
it to t4g.medium.