Fix azure publishing (#308)
azure publishing was failing with a usage message of azure-image-upload
saying 'Got 4 args, expected 1'

The '4' in that string came from 'names' being expanded on by shell and
sent to parallel on a single line rather than as 4 lines. The result was
that image-upload was called once with 4 args rather than 4 times with 1
arg.

Other changes here:
* do not quote name={}. parallel seems to try to be shell escaping. the
result was that name was actually getting literal "azure-base"
(including the quotes)
* catch error if 'git rev-parse HEAD' errors rather than silently using
empty value.
 * print the message with 'names' on a single line rather than N lines.
* change passing of parameters to --name=value rather than --name value.
* use --no-run-if-empty to avoid running with empty value due to a
trailing newline