nvidia-cuda-ld-config: Add a -dev variant (#2459)
libcuda.so.1 and other files from the host are normally injected into the
container at runtime because they are tightly coupled with the version
of the GPU driver running on the host.

NVIDIA provides stub libraries to use for build-time linking, and
these -dev packages will add them to the library search path for
that purpose. It is also useful at test time, so lddcheck will not
report unresolved libraries. melange sca already ignores libcuda.so.1
when scanning for dependencies.

Signed-off-by: dann frazier <dann.frazier@chainguard.dev>