Adjust ssh to utilize ssh agent and default key to ~/.ssh/id_ed25519
Previously you had to provide --private-key to 'ssh' of
gcp, qemu, azure, aws.  That was annoying.
Now it does the right thing and uses the ssh agent if present
and picks id_ed25519 key in ~/.ssh if that is present.

Also here are changes to qemu to add all public keys in start.