Do not require a public key on launch for aws, azure or gcp.
Instead of requiring a public key, pick one.

The default behavior will be to get ~/.ssh/id_ed25519.pub
if it is present, then id_rsa.pub and then the keys listed
in ssh agent (`ssh-add -L` will show those).

This can obviously be incorrect. In that case the caller should
provide a --private-key.