Support WVM_DISPLAY and WVM_VNC variables in 'make run-<target>' (#337)
This is useful to allow you to get a vnc or sdl console. To demonstrate:

    $ WVM_DISPLAY=sdl WVM_VNC=localhost:0 make run-debug-generic
    ...

    Starting Record Runlevel Change in UTMP...
    [  OK  ] Finished Record Runlevel Change in UTMP.
    Chainguard
    Kernel 6.14.6-r2-linux on an x86_64 (ttyS0)

    chainguard login:

This will pop up the "normal" qemu sdl window.
It will also allow you to attach to localhost:0 with your favorite vnc
client.

Also note that while the above is running, you can

    make debug-shell-generic

And it will drop you into a root prompt.