helpers/test-wolfi-vm*: fix test upload to work with test groups (#58)
Fix the runners for QEMU, AWS, and GCP to upload tests again after the
introduction of test groups. For now this restores the previous behavior of
iterating over all tests of all groups for those runners which was broken and
resulting in this error:

    2025/05/15 20:06:13 failed to move [...]/vms-test/test/x86_64/core to remote host: \
                        failed to shove file: read [...]/vms-test/test/x86_64/core: is a directory

Leave the runner for Azure untouched since it has been modified for the purpose
of introducing test groups already.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>