vms-test: upload test artifacts regardless of outcome (#346)
By default, every step in a GitHub test action is executed only if the
action has not failed yet. This causes the vms-test action to only
upload test artifacts if the test execution succeeded, and to discard
the artifacts when they are most needed. Modify the vms-test action to
execute the upload step regardless of the action's success state.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>