helpers/test-wolfi-vm*: make all runners only collect metrics if there are any
In 85478ae5 ("helpers/test-wolfi-vm: implement test configuration v1 parsing
for Azure") the runner for Azure was enhanced to only attempt collecting
metrics if the current test emits any. Bring this improvement to the QEMU, AWS,
and GCP runners as well, so that purely functional tests without metrics can be
run via them, too.

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>