AWS: add capability to launch instances with user data (#61)
Add the ability to launch instances with user data [1] to the AWS test runner.
Since the test runner reuses instances for multiple test runs, all instances
will have to share a user data file for now. Should this become a problem in
the future, the runner will have to launch separate instances for tests with
dedicated user data.

[1]: https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/user-data.html

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>