fix(emptypackage.yaml): The empty package test should fail if the package is not empty (#23931)
Our understanding of the empty package test was that he wanted to
detect packages which are expected to be empty and are no long empty.

If the test is not exiting 1 then we will never know if the package is no longer empty.

This follows on from corresponding wolfi-dev/os PR @  https://github.com/wolfi-dev/os/pull/54029

Signed-off-by: philroche <phil.roche@chainguard.dev>