openssl-config-fipshardened: temporarily stop using P521 for openssh-server
Commit f4ea45a9 ("openssl-config-fipshardened: temporarily stop using P521")
dropped algorithms based on the NIST P-521 curve from the hardened
openssh-client config due to [1]. Do the same for openssh-server now that
there's a dedicated config for it.

Note: This intentionally leaves openssl-config-fipshardened-3.4.0 alone, as our
build of OpenSSL 3.4.0 carries the cherry-picks needed to fix the CVE we do
this for. [2]

Note 2: Intentionally not bumping the epoch. This change is split out of the
previous commit for better tracing only. Both are pushed together (and the
previous commit does bump the epoch).

[1]: https://openssl-library.org/news/secadv/20250120.txt
[2]: https://github.com/chainguard-dev/openssl/pull/13

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>