ct-manylinux-2.28: relax dependency on major gcc version
Currently released version of crosstool-ng only supports gcc-14 and
enablement of gcc-15 is in-progress upstream. Thus unless crosstool-ng
snapshot is packaged, we cannot currently upgrade ct-manylinux-2.28 to
gcc-15.

Separately, we do build and provide libstdc++ headers, which currently
come from gcc-14 release.

It is not clear to me if it is allowed to use g++-15 with
libstdc++-14, howevever it would seem to be safe because headers match
the link-time library.

This helps to resolve the package installation by melange.