Pin aws-eks-* variants to Amazon EKS 1.32 (#391)
Since we did not pin any relevant packages, AMIs published within the
last week have picked up parts of Kubernetes 1.33 as Amazon EKS 1.33
became available. We made the decision to pin to Kubernetes 1.32 since
this is what the Marketplace
listings have been targeting since their creation. We will publish
images targeting Kubernetes 1.33 soon, but with a dedicated Marketplace
listing.

Rename the Amazon EKS variants to include the version of Kubernetes they
are targeting. Retain the existing product listings, but update the AMI
names and descriptions to reflect they are targeting a particular
version of Kubernetes.

---------

Signed-off-by: Markus Boehme <markus.boehme@chainguard.dev>